requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.2.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.0
fastapi-cli==0.0.5
greenlet==3.1.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
psycopg2==2.9.9
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pytest==8.3.3
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.2
rich==13.8.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.5
typer==0.12.5
typing_extensions==4.12.2
uvicorn==0.30.6
watchfiles==0.24.0
websockets==13.0.1
